feat(miniapp): extract constants to miniapp-builder-shared
