+++ packages/miniapp-builder-shared/README.md
@@ -1,12 +1,15 @@
 ### Shared lib for miniapp builder
 
 #### filterNativePages
+
 It will filter native miniapp pages and return a need copy native page files list.
 
 #### getAppConfig
+
 It will return a vaild app config for miniapp.
 
 #### pathHelper
+
 It will export the following methods:
   - `relativeModuleResolve`
   - `normalizeOutputFilePath`
@@ -20,4 +23,9 @@ It will export the following methods:
   - `getBundlePath`
 
 #### platformMap
+
 It will return miniapp platform and suffix name correspondence.
+
+#### constants
+
+It will return miniapp related constants
