+++ packages/miniapp-builder-shared/src/index.js
@@ -2,10 +2,12 @@ const filterNativePages = require('./filterNativePages');
 const getAppConfig = require('./getAppConfig');
 const pathHelper = require('./pathHelper');
 const platformMap = require('./platformMap');
+const constants = require('./constants');
 
 module.exports = {
   filterNativePages,
   getAppConfig,
   pathHelper,
-  platformMap
+  platformMap,
+  constants
 };
